requirements.txt
fastapi
uvicorn
supabase
python-dotenv
gunicorn
slowapi